+++ instrumentation-genai/opentelemetry-instrumentation-crewai/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Version 0.1.2 - 2026-01-28
+
+### Changed
+- **Fixed input and output messages** - Workflow and AgentInvocation's input and output messages are serialized per item.
+
 ## [0.1.1] - 2026-01-26
 
 ### Changed