+++ instrumentation-genai/opentelemetry-instrumentation-langchain/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this repository are documented in this file.
 
+## Version 0.1.6 - 2026-01-28
+
+### Changed
+- **Fixed input and output messages** - Workflow and AgentInvocation's input and output messages are serialized per item.
+
 ## Version 0.1.5 - 2026-01-27
 
 ### Changed